// src/dragonball/models/Tags.ts
export interface Tags {
  name: string;
  description: string;
  sex: string;
  birthday: string;
  height: string;
  weight: string;
  alias: string;
  race: string;
  death: string;
  address: string;
  occupation: string;
  relatives: string;
  photo: string;
  appears: string;
  transformation: string;
}